Fix: made easy path in links, create directories for each page
