+++ src/pages/constructor/constructor.jsx
@@ -4,10 +4,10 @@ import { useSelector } from 'react-redux';
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 import styles from './constructor.module.css';
-import AppHeader from '../components/app-header/app-header';
-import BurgerIngredients from '../components/burger-ingredients/burger-ingredients';
-import BurgerConstructor from '../components/burger-constructor/burger-constructor';
-import Modal from '../components/modal/modal';
+import AppHeader from '../../components/app-header/app-header';
+import BurgerIngredients from '../../components/burger-ingredients/burger-ingredients';
+import BurgerConstructor from '../../components/burger-constructor/burger-constructor';
+import Modal from '../../components/modal/modal';
 
 export const ConstructorPage = () => {
     const loading = useSelector(store => store.loading);
