+++ src/pages/detail/detail.module.css
