+++ src/pages/index.js
@@ -1,11 +1,11 @@
-import { ConstructorPage } from "./constructor";
-import { LoginPage } from "./login";
-import { ProfilePage } from "./profile";
-import { ForgotPassPage } from "./forgotPass";
-import { RegistrationPage } from "./registration";
-import { ResetPassPage } from "./resetPass";
-import { IngredientDetailPage } from "./detail";
-import { NotFoundPage } from "./404";
+import { ConstructorPage } from "./constructor/constructor";
+import { LoginPage } from "./login/login";
+import { ProfilePage } from "./profile/profile";
+import { ForgotPassPage } from "./forgotPass/forgotPass";
+import { RegistrationPage } from "./registration/registration";
+import { ResetPassPage } from "./resetPass/resetPass";
+import { IngredientDetailPage } from "./detail/detail";
+import { NotFoundPage } from "./404/404";
 
 export {
     ConstructorPage,
