+++ src/pages/resetPass/resetPass.module.css
